requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pandas==2.3.3
numpy==2.2.5
scipy==1.14.1
scikit-learn==1.5.2
matplotlib==3.9.2
pydantic==2.8.2
jinja2==3.1.4
sqlalchemy==2.0.32
openai>=1.0.0
python-dotenv==1.0.0